Ensure values passed to fontSizeToPx are strings
Closes #121.

There are occasions where, when no units are present, a value is
passed along as an integer by cssstats. This is ultimately something that
should be handled by cssstats, to ensure that it is passing around
the intended object types. However, it never hurts to ensure an object
is a string before calling match.